+++ packages/core/src/components/ozon-content/components/ozon-content-toggletip/ozon-content-toggletip.tsx
@@ -5,6 +5,10 @@ import { Tooltip } from "../../../../functional-components/tooltip/tooltip.funct
 import { TooltipClean } from "../../../../functional-components/tooltip/tooltip.interfaces";
 import { IconAlias } from "../../../icon/icon.interfaces";
 
+/**
+ * @slot - A slot to place the main content inside the toggletip overlay.
+ * @slot label - A slot to place the trigger label text of the toggletip.
+ */
 @Component({
   tag: "dso-ozon-content-toggletip",
   styleUrl: "./ozon-content-toggletip.scss",