+++ packages/dso-toolkit/src/components/plekinfo-card/plekinfo-card.args.ts
@@ -16,6 +16,7 @@ export interface PlekinfoCardArgs {
   interaction?: SlideToggle;
   active: boolean;
   meta: Label;
+  showStroke: boolean;
   wijzigactie: PlekinfoWijzigactie;
   dsoPlekinfoCardClick: HandlerFunction;
 }
@@ -26,6 +27,7 @@ export const plekinfoCardArgs: Omit<PlekinfoCardArgs, "meta" | "wijzigactie"> = 
   interaction: undefined,
   label: "Radarverstorende bouwwerken",
   active: false,
+  showStroke: true,
   dsoPlekinfoCardClick: fn(),
 };
 
@@ -47,6 +49,11 @@ export const plekinfoCardArgTypes: ArgTypes<Omit<PlekinfoCardArgs, "meta">> = {
     },
   },
   active: {
+    control: {
+      type: "boolean",
+    },
+  },
+  showStroke: {
     control: {
       type: "boolean",
     },
@@ -73,6 +80,7 @@ export function plekinfoCardArgsMapper<TemplateFnReturnType>(
     href: a.href,
     targetBlank: a.targetBlank,
     active: a.active,
+    showStroke: a.showStroke,
     wijzigactie: a.wijzigactie,
     interaction: a.interaction,
     meta: a.meta,