+++ CHANGELOG.md
@@ -9,6 +9,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 * Accordion: Optionele Badge in handle van Accordion Section ([#3936](https://github.com/dso-toolkit/dso-toolkit/issues/3936))
 * Header: Help-knop in menu bij mobiele viewport ([#3017](https://github.com/dso-toolkit/dso-toolkit/issues/3017))
 
+### Task
+* Packages: Dependencies updates ([#3937](https://github.com/dso-toolkit/dso-toolkit/issues/3937))
+
 ## 🧧 Release 100.3.1 - 2026-09-08
 
 ### Fixed